# Optimizing CompareArrays