--- a/src/Medo.Uuid7/Uuid7.cs
+++ b/src/Medo.Uuid7/Uuid7.cs
@@ -1071,7 +1071,18 @@ public readonly struct Uuid7
 
 #if NET6_0_OR_GREATER
     [MethodImpl(MethodImplOptions.AggressiveOptimization)]
-#endif
+    private static int CompareArrays(ReadOnlySpan<byte> buffer1, ReadOnlySpan<byte> buffer2) {
+        if (buffer1.Length == 16 && buffer2.Length == 16) {
+            return buffer1.SequenceCompareTo(buffer2);
+        } else if (buffer1.Length != 16) {
+            return -1;
+        } else if (buffer2.Length != 16) {
+            return +1;
+        }
+
+        return 0;  // object are equal
+    }
+#else
     private static int CompareArrays(byte[] buffer1, byte[] buffer2) {
         if ((buffer1 != null) && (buffer2 != null) && (buffer1.Length == 16) && (buffer2.Length == 16)) {  // protecting against EF or similar API that uses reflection (https://github.com/medo64/Medo.Uuid7/issues/1)
             var comparer = Comparer<byte>.Default;
@@ -1087,7 +1098,7 @@ public readonly struct Uuid7
 
         return 0;  // object are equal
     }
-
+#endif
 
     private static readonly RandomNumberGenerator Random = RandomNumberGenerator.Create();  // needed due to .NET Standard 2.0
 #if !UUID7_NO_RANDOM_BUFFER