This change improves the performance of the internal `CompareArrays` method by delegating most of the work to the heavily-optimized [SequenceCompareTo](https://learn.microsoft.com/en-us/dotnet/api/system.memoryextensions.sequencecompareto?view=net-8.0) method.

On my system, this is 30-40% faster, but results will likely vary depending on the SIMD capabilities of the underlying processor. See the new CompareArrays benchmark for details.

```
BenchmarkDotNet v0.13.6, macOS Sonoma 14.2.1 (23C71) [Darwin 23.2.0]
Apple M1 Pro, 1 CPU, 10 logical and 10 physical cores
.NET SDK 8.0.101
  [Host]     : .NET 8.0.1 (8.0.123.58001), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 8.0.1 (8.0.123.58001), Arm64 RyuJIT AdvSIMD
```

|       Method | Iterations |         Mean |     Error |    StdDev | Ratio |
|------------- |----------- |-------------:|----------:|----------:|------:|
|     Original |        100 |   1,062.6 ns |   2.40 ns |   2.13 ns |  1.00 |
|       Switch |        100 |   1,056.5 ns |   1.72 ns |   1.53 ns |  0.99 |
|   SwitchSpan |        100 |   1,902.3 ns |   2.19 ns |   1.94 ns |  1.79 |
| SpanSequence |        100 |     708.4 ns |   0.97 ns |   0.86 ns |  0.67 |
|              |            |              |           |           |       |
|     Original |       1000 |  10,281.5 ns |   4.92 ns |   4.10 ns |  1.00 |
|       Switch |       1000 |  10,284.7 ns |   3.96 ns |   3.31 ns |  1.00 |
|   SwitchSpan |       1000 |  18,508.2 ns |  39.03 ns |  32.59 ns |  1.80 |
| SpanSequence |       1000 |   7,229.9 ns |   5.98 ns |   5.59 ns |  0.70 |
|              |            |              |           |           |       |
|     Original |      10000 |  96,331.5 ns | 114.37 ns | 101.39 ns |  1.00 |
|       Switch |      10000 |  99,886.0 ns | 129.02 ns | 114.38 ns |  1.04 |
|   SwitchSpan |      10000 | 170,427.8 ns | 177.83 ns | 157.64 ns |  1.77 |
| SpanSequence |      10000 |  66,711.5 ns | 217.44 ns | 203.40 ns |  0.69 |