@jtmueller 2024-01-15
I added an optimization to `TryParseAsString` that was present in similar methods, but I somehow missed before.