@jtmueller 2024-01-15
On a different topic, it looks like using utf8-byte-based alphabet dictionaries speeds up dictionary lookup by 1-3%. This isn't enough of a benefit to really be worth bothering with, except that it might facilitate implementing [IUtf8SpanParseable<Uuid7>](https://learn.microsoft.com/en-us/dotnet/api/system.iutf8spanparsable-1?view=net-8.0) so I thought I'd throw it out there as a suggestion.

|   Method |     Mean |   Error |  StdDev | Ratio |
|--------- |---------:|--------:|--------:|------:|
| Original | 110.9 ns | 0.78 ns | 0.65 ns |  1.00 |
| Utf8Dict | 108.4 ns | 0.61 ns | 0.54 ns |  0.98 |

```csharp
namespace Uuid7Benchmark;
using System;
using System.Numerics;
using System.Collections.Generic;
using System.Runtime.CompilerServices;
using System.Security.Cryptography;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using Medo;

public class Utf8Alphabet
{
    [Benchmark(Baseline = true)]
    public void Original()
    {
        for (var i = 0; i < Base58Alphabet.Length; i++)
        {
            _ = _base58Lookup[Base58Alphabet[i]];
        }
    }

    [Benchmark]
    public void Utf8Dict()
    {
        for (var i = 0; i < Base58Alphabet.Length; i++)
        {
            _ = _base58Lookup_Utf8[(byte)Base58Alphabet[i]];
        }
    }

    private static readonly char[] Base58Alphabet = [
        '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A',
        'B', 'C', 'D', 'E', 'F', 'G', 'H', 'J', 'K', 'L',
        'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W',
        'X', 'Y', 'Z', 'a', 'b', 'c', 'd', 'e', 'f', 'g',
        'h', 'i', 'j', 'k', 'm', 'n', 'o', 'p', 'q', 'r',
        's', 't', 'u', 'v', 'w', 'x', 'y', 'z'
    ];

    // this is optimized by the C# compiler into an embedded resource
    private static ReadOnlySpan<byte> Base58Alphabet_Utf8 => "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz"u8;

    private Dictionary<char, BigInteger> _base58Lookup = new();
    private Dictionary<byte, BigInteger> _base58Lookup_Utf8 = new();

    [GlobalSetup]
    public void Setup()
    {
        _base58Lookup = new Dictionary<char, BigInteger>(Base58Alphabet.Length);
        for (var i = 0; i < Base58Alphabet.Length; i++)
        {
            _base58Lookup.Add(Base58Alphabet[i], i);
        }

        _base58Lookup_Utf8 = new Dictionary<byte, BigInteger>(Base58Alphabet.Length);
        for (var i = 0; i < Base58Alphabet.Length; i++)
        {
            _base58Lookup_Utf8.Add(Base58Alphabet_Utf8[i], i);
        }
    }
}
```