@medo64 2024-01-16
Looks good. Thank you!
Merged in. Will check it further over weekend and probably push new version then.