@medo64 2024-01-16
As for tf8-byte-based alphabet dictionaries, I am good with either, assuming it works for .NET Standard 2.0 and doesn't interfere with UTF-8 parsing. Tentatively, I would say it's ok for both but I'll look into it futher.

Thank you for going over code and getting all these improvements!